requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==3.1
django-crispy-forms==1.9.2
django-json-field==0.5.7
django-jsonfield==1.4.0
django-phone-field==1.8.1
djangorestframework==3.11.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jsonfield==3.1.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
Werkzeug==1.0.1
